Eliminar, actualizar, sql

+++ Zoo/Zoo/Models/Clasificaciones.cs
@@ -8,7 +8,7 @@ namespace Zoo
     public class Clasificaciones
     {
 
-        public long idClasificacion { get; set; }
+        public int idClasificacion { get; set; }
         public string denominacion { get; set; }
     }
 }
